Prints a completeness report for a country or admin1 unit

This function is designed to give an idea of the completeness of the HIT-COVID database for a specified country or administrative level 1 unit. It prints information about the date the location was previously updated, what percentage of the entries have been verified by contributors, and the status of each intervention group (complete, incomplete, unsure).

check_completeness(hit_data, country = NULL, admin1 = NULL)

Arguments

hit_data
the full HIT-COVID database pulled from GitHub, pulled using hit_pull

country
country ISO 3166-1 alpha-3 code

admin1
vector of the first administrative unit GID codes to filter the data to (see geo_lookup for concordance of admin 1 codes to names).

Value

The function does not return an object but prints the result in the console.

Details

The HIT-COVID database must first be loaded with hit_pull and fed into this function. The report can only be request for one country or admin1 unit. If a country and admin1 unit are specified, the country input will be ignored and the report for the admin1 unit will be printed. The admin1 units that are included in the HIT-COVID database can be found by using get_admin1.

The completeness table which prints at the end of the report gives the completeness status of each intervention group. This status is recorded by contributors (usually the contributors who logged the data). The column "Last Update" is the date that intervention group was last updated, if this column is NA then this intervention group was never updated. The column "Last Marked Complete" is the date that the intervention group was last recorded as complete by a contributor. The column "Current Status" is the most recent status for that intervention group recorded by a contributor and "Date of Current Status" is the date that most recent status was recorded.

An important note is that when requesting a national level report, the completeness table only reflects national-level entries. If fewer than 20% of all entries for that country are at the national level, the function gives a message that this table may be misleading. That is because for those countries the interventions are likely being managed at the admin1 level and therefore reports for specific admin1 units might be more useful.

Examples

# Puling the HIT-COVID database
hit_data <- hit_pull(add_first_case = FALSE)

# Printing report for Zimbabwe
check_completeness(hit_data, country = "ZWE")

#> 
#> Completeness Report for Zimbabwe (ZWE).
#> 
#> There have been 49 national level policy changes logged.
#> There have been 0 admin1 level policy changes logged.
#> Therefore, 100% of the updates are at the national level.
#> 
#> Of all of the entries for this country, 49 (100%) were marked as verified.
#> At the national level, the last day that a intervention update or 'no updates' was logged was: 2020-11-04.
#> Table of Intervention Completeness
#> 
#> |Intervention                        |Last Update |Last Marked Complete |Current Status |Date of Current Status |
#> |:-----------------------------------|:-----------|:--------------------|:--------------|:----------------------|
#> |Border closures                     |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Contact tracing                     |2020-04-02  |NA                   |Unsure         |2020-11-13             |
#> |Household confinement               |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Leisure and entertainment closures  |2020-04-02  |2020-05-07           |Unsure         |2020-11-13             |
#> |Limiting movement within borders    |2020-11-04  |NA                   |Unsure         |2020-11-13             |
#> |Limiting size of gatherings         |2020-04-02  |2020-05-07           |Unsure         |2020-11-13             |
#> |Long-term care closures             |NA          |NA                   |Unsure         |2020-11-13             |
#> |Military and police deployment      |2020-04-02  |2020-05-07           |Unsure         |2020-11-13             |
#> |Office closures                     |2020-11-04  |NA                   |Unsure         |2020-11-13             |
#> |Public space closures               |NA          |2020-05-07           |Unsure         |2020-11-13             |
#> |Public transportation closures      |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Quarantine and home isolation       |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Religious venue closures            |NA          |NA                   |Unsure         |2020-11-13             |
#> |Restaurant closures                 |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Retail store closures               |2020-11-04  |NA                   |Unsure         |2020-11-13             |
#> |School closures                     |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |State of emergency                  |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Symptom screening                   |2020-11-04  |2020-05-07           |Unsure         |2020-11-13             |
#> |Testing of asymptomatic individuals |NA          |NA                   |Unsure         |2020-11-13             |
#> |Testing of symptomatic individuals  |2020-04-02  |2020-05-07           |Unsure         |2020-11-13             |
#> |Universal facemask policies         |2020-11-04  |NA                   |Unsure         |2020-11-13             |

# Printing report for Great Britain
check_completeness(hit_data, country = "GBR")

#> 
#> Completeness Report for United Kingdom (GBR).
#> 
#> There have been 11 national level policy changes logged.
#> There have been 161 admin1 level policy changes logged.
#> Therefore, 6.4% of the updates are at the national level.
#> 
#> Of all of the entries for this country, 135 (78.5%) were marked as verified.
#> At the national level, the last day that a intervention update or 'no updates' was logged was: 2020-06-08.
#> At the admin1 level, the last day that a intervention update or 'no updates' was logged was: 2020-11-18.
#> 
#> Table of Intervention Completeness
#> 
#> |Intervention                        |Last Update |Last Marked Complete |Current Status |Date of Current Status |
#> |:-----------------------------------|:-----------|:--------------------|:--------------|:----------------------|
#> |Border closures                     |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Contact tracing                     |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Household confinement               |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Leisure and entertainment closures  |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Limiting movement within borders    |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Limiting size of gatherings         |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Long-term care closures             |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Military and police deployment      |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Office closures                     |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Public space closures               |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Public transportation closures      |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Quarantine and home isolation       |2020-06-08  |2020-09-26           |Complete       |2020-09-26             |
#> |Religious venue closures            |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Restaurant closures                 |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Retail store closures               |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |School closures                     |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |State of emergency                  |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Symptom screening                   |NA          |2020-09-26           |Complete       |2020-09-26             |
#> |Testing of asymptomatic individuals |2020-04-30  |2020-09-26           |Complete       |2020-09-26             |
#> |Testing of symptomatic individuals  |2020-05-19  |2020-09-26           |Complete       |2020-09-26             |
#> |Universal facemask policies         |2020-05-12  |2020-09-26           |Complete       |2020-09-26             |

#> Fewer than 20% of the updates for this country are at the national level so this table
#>  may be misleading. You can run this function at the admin1 level for more detailed information.

# Printing report for England
check_completeness(hit_data, admin1 = "GBR.1_1")

#> 
#> Completeness Report for England (GBR.1_1).
#> 
#> There have been 11 national level policy changes logged.
#> There have been 57 admin1 level policy changes logged.
#> Of all of the entries for this country, 50 (87.7%) were marked as verified.
#> At the national level, the last day that a intervention update or 'no updates' was logged was: 2020-06-08.
#> At the admin1 level, the last day that a intervention update or 'no updates' was logged was: 2020-11-05.
#> 
#> Table of Intervention Completeness
#> 
#> |Intervention                        |Last Update |Last Marked Complete |Current Status |Date of Current Status |
#> |:-----------------------------------|:-----------|:--------------------|:--------------|:----------------------|
#> |Border closures                     |2020-07-14  |2020-10-03           |Complete       |2020-10-03             |
#> |Contact tracing                     |2020-09-11  |2020-10-03           |Complete       |2020-10-03             |
#> |Household confinement               |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Leisure and entertainment closures  |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Limiting movement within borders    |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Limiting size of gatherings         |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Long-term care closures             |NA          |2020-10-03           |Complete       |2020-10-03             |
#> |Military and police deployment      |NA          |2020-10-03           |Complete       |2020-10-03             |
#> |Office closures                     |2020-03-29  |2020-10-03           |Complete       |2020-10-03             |
#> |Public space closures               |2020-07-14  |2020-10-03           |Complete       |2020-10-03             |
#> |Public transportation closures      |2020-03-30  |2020-10-03           |Complete       |2020-10-03             |
#> |Quarantine and home isolation       |2020-09-29  |2020-10-03           |Complete       |2020-10-03             |
#> |Religious venue closures            |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Restaurant closures                 |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |Retail store closures               |2020-11-05  |2020-10-03           |Complete       |2020-10-03             |
#> |School closures                     |2020-09-09  |2020-10-03           |Complete       |2020-10-03             |
#> |State of emergency                  |NA          |2020-10-03           |Complete       |2020-10-03             |
#> |Symptom screening                   |NA          |2020-10-03           |Complete       |2020-10-03             |
#> |Testing of asymptomatic individuals |2020-06-10  |2020-10-03           |Complete       |2020-10-03             |
#> |Testing of symptomatic individuals  |2020-04-30  |2020-10-03           |Complete       |2020-10-03             |
#> |Universal facemask policies         |2020-09-29  |2020-10-03           |Complete       |2020-10-03             |