Pulls HIT-COVID database from GitHub

This function pulls the HIT-COVID database from the GitHub repo: https://github.com/HopkinsIDD/hit-covid. If desired, it also uses the covidregionaldata package to add the date of the first case of COVID-19 and the date of the first death from COVID-19. The source is either the European Centre for Disease Control (ECDC) or the World Health Organization (WHO).

hit_pull(add_first_case = TRUE, source = c("WHO", "ECDC"))

Arguments

add_first_case
a logical indicating if information about the first case and the first death for each country should be added to the dataset

source
the source of the case data that is used to determine the date of first case and first death if add_first_case is TRUE: one of "ECDC" or "WHO" (default is "WHO")

Value

A dataframe with columns as described in the GitHub README (https://github.com/HopkinsIDD/hit-covid)

References

Sam Abbott, Katharine Sherratt, Jonnie Bevan, Hamish Gibbs, Joel Hellewell, James Munday, Paul Campbell and Sebastian Funk (2020). covidregionaldata: Subnational Data for the Covid-19 Outbreak. R package version 0.6.0. https://CRAN.R-project.org/package=covidregionaldata

See also

Examples

# \donttest{
#Pulling the HIT-COVID database and adding date of first case and death from WHO
hit_data <- hit_pull(add_first_case = TRUE, source = "WHO")

#> Error in hit_pull(add_first_case = TRUE, source = "WHO"): could not find function "hit_pull"

#Pulling the HIT-COVID databasea and not adding the date of first case and first death
hit_data <- hit_pull(add_first_case = FALSE)

#> Error in hit_pull(add_first_case = FALSE): could not find function "hit_pull"

# }